+++ Changelog.md
@@ -3,6 +3,7 @@ Version 0.1.0 (Unreleased)
 
 Features
 --------
+- Encode logs as JSON objects.
 - Implement HTTP headers injection/extraction.
 - Implement thrift-based binary injection/extraction.
 - Initial library structure.